+++ server/src/modules/infrastructure/exchange-init/exchange-init.module.ts
@@ -1,9 +1,11 @@
 import { Module, Global } from '@nestjs/common';
+import { CacheModule } from '@nestjs/cache-manager';
 import { ExchangeInitService } from './exchange-init.service';
 
 @Global()
 @Module({
+  imports: [CacheModule.register()],
   providers: [ExchangeInitService],
   exports: [ExchangeInitService],
 })
-export class ExchangeInitModule { }
+export class ExchangeInitModule {}